feat(classroom-app): Ajuste al dia de ayer

+++ lib/features/accounts/domain/services/calculate_credit_amount_service.dart
@@ -1,6 +1,6 @@
-double calculateCreditAmount({required int score, required int payCapacity}) {
+int calculateCreditAmount({required int score, required int payCapacity}) {
   if (score > 700 && payCapacity > 50000) {
-    return (payCapacity * 0.2); // Monto aprobado
+    return (payCapacity * 0.2).toInt(); // Monto aprobado
   } else {
     return 0; // Crédito denegado
   }
